Remove some legacy cruft
The latest versions of flake8 and flake8-import-order can be used just
fine now (they've been ok for some time).

Since with google style flake8-import-order would generate more issues
than with the cryptography style I decided to switch and fix one thing
it complained about.

We switch to pycodestyle instead of pinned pep8 version as that pep8
version can't be installed with latest flake8 and the name of the
package has been changed to pycodestyle. We still pin the version though
as there's a bad interaction between the latest pycodestyle and the
latest flake8.