Reformat the code using Black
We could use some style consistency in the project and Black looks like
the best tool for the job.

Two flake8 errors are being silenced from now on:

* E203 whitespace before :
* W503 line break before binary operator

Both are to satisfy Black-formatted code (and W503 is somemwhat against
the latest PEP8 recommendations regarding line breaks and binary
operators in new code).