First stab at supporting listening on IPv6 interfaces
This change adds basic support for listening on IPv6 interfaces.
Some limitations exist for non-POSIX platforms, pending fixes in
Python and in the ifaddr library. Also dual V4-V6 sockets may not
work on all BSD platforms. As a result, V4-only is used by default.

Unfortunately, Travis does not seem to support IPv6, so the tests
are disabled on it, which also leads to coverage decrease.