Autorefresh the list of network interfaces
Basically create_sockets has been renamed refresh_sockets, and moved
into the Zeroconf class. It can now check for changes and only refresh
if needed.  We poll every 3 minutes from within the engine loop to
detect changes.

In addition, when we get an error in send_from(), we assume that it's
due to network changes, refresh the sockets, and retry with all
new sockets if we were multicasting something.