Dont recalculate the expiration and stale time every update
I have a network with 170 devices running Zeroconf. Every minute a zeroconf request for broadcast is cast out. Then we were listening for Zeroconf devices on that network.
To get a more realistic test, the Zeroconf ServiceBrowser is ran on a separate thread from a main thread. On the main thread an I/O limited call to QNetworkManager is made every 2 seconds,
in order to include performance penalties due to thread switching. The experiment was ran on a MacBook Air 2015 (Intel Core i7-5650U) through Ubuntu 19.10 and Python 3.7.
This was left running for exactly one hour, both before and after this commit.

Before this commit, there were 132107499 (132M) calls to the get_expiration_time function, totalling 141.647s (just over 2 minutes).
After this commit, there were 1661203 (1.6M) calls to the get_expiration_time function, totalling 2.068s.

This saved about 2 minutes of processing time out of the total 60 minutes, on average 3.88% processing power on the tested CPU. It is expected to see similar improvements on all CPU architectures.