Provide and enforce type hints everywhere except for tests
The tests' time will come too in the future, though, I think. I believe
nose has problems with running annotated tests right now so let's leave
it for later.

DNSEntry.to_string renamed to entry_to_string because DNSRecord
subclasses DNSEntry and overrides to_string with a different signature,
so just to be explicit and obvious here I renamed it – I don't think any
client code will break because of this.

I got rid of ServicePropertiesType in favor of generic Dict because
having to type all the properties got annoying when variance got
involved – maybe it'll be restored in the future but it seems like too
much hassle now.