Ensure the name cache is rolled back when the packet reaches maximum size
If the packet was too large, it would be rolled back at the end of write_record.
We need to remove the names that were added to the name cache (self.names)
as well to avoid a case were we would create a pointer to a name that was
rolled back.

The size of the packet was incorrect at the end after the inserts because
insert_short would increase self.size even though it was already accounted
before. To resolve this insert_short_at_start was added which does not
increase self.size. This did not cause an actual bug, however it sure
made debugging this problem far more difficult.

Additionally the size now inserted and then replaced when the actual
size is known because it made debugging quite difficult since the size
did not previously agree with the data.