RecordUpdateListener now uses update_records instead of update_record
This allows the listener to receive all the records that have
been updated in a single transaction such as a packet or
cache expiry.

`update_record` has been deprecated in favor of `update_records`
A compatibility shim exists to ensure classes that use
`RecordUpdateListener` as a base class continue to have
`update_record` called, however they should be updated
as soon as possible.

A new method `update_records_complete` is now called on each
listener when all listeners have completed processing updates
and the cache has been updated. This allows ServiceBrowsers
to delay calling handlers until they are sure the cache
has been updated as its a common pattern to call for
ServiceInfo when a ServiceBrowser handler fires.