Qualify IPv6 link-local addresses with scope_id (#343)
Co-authored-by: Lokesh Prajapati <lokesh.prajapati@ncipher.com>
Co-authored-by: de Angelis, Antonio <Antonio.deAngelis@ncipher.com>

When a service is advertised on an IPv6 address where
the scope is link local, i.e. fe80::/64 (see RFC 4007)
the resolved IPv6 address must be extended with the
scope_id that identifies through the "%" symbol the
local interface to be used when routing to that address.
A new API `parsed_scoped_addresses()` is provided to
return qualified addresses to avoid breaking compatibility
on the existing parsed_addresses().