New ServiceBrowsers now request QU in the first outgoing when unspecified
https://datatracker.ietf.org/doc/html/rfc6762#section-5.4

When we start a ServiceBrowser and zeroconf has just started up, the known
answer list will be small. By asking a QU question first, it is likely
that we have a large known answer list by the time we ask the QM question
a second later (current default). This reduces the amount of traffic on
the network, and has the secondary advantage that most responders will
answer a QU question without the typical delay answering QM questions.