fix: reduce denial of service protection overhead
If there is a device with a bad mdns stack that is flooding
the network with response packets there is little we can do
but optimize how we handle this as mdns works on a trust model
of generally accepting network traffic. Its unlikely that the
user is going to remove the offending the device so we will
reduce the overhead to deal with one of these.