feat: ensure ServiceInfo.properties always returns bytes
Clean up the typing on `ServiceInfo.properties` to always return bytes.

The typing on properties was very confusing because it could have a mix of bytes
and str types, but only bytes would ever come back from the backend and cache.
The only way str could happen is if someone manually passed it in.

We now ensure all passed in data is converted to bytes