Use per-interface per-protocol sockets
Instead of using one global socket for all interfaces, use one socket
per interface and per protocol. This allows better control over
which interfaces are used for multicast and unicast, and allows
to know which interface a packet was received on.

TODO: Uses Linux/macOS specific socket options to bind to an interface.